+++ lib/input/read_until.go
@@ -57,7 +57,7 @@ input:
 				"A [Bloblang query](/docs/guides/bloblang/about/) that should return a boolean value indicating whether the input should now be closed.",
 				`this.type == "foo"`,
 				`count("messages") >= 100`,
-			).HasDefault(""),
+			).HasDefault("").Linter(docs.LintBloblangMapping),
 			docs.FieldDeprecated("condition").HasType(docs.FieldCondition).OmitWhen(func(field, _ interface{}) (string, bool) {
 				defaultBytes, err := yaml.Marshal(condition.NewConfig())
 				if err != nil {