+++ lib/message/batch/docs.go
@@ -42,7 +42,7 @@ Allows you to configure a [batching policy](/docs/configuration/batching).`,
 				"check",
 				"A [Bloblang query](/docs/guides/bloblang/about/) that should return a boolean value indicating whether a message should end a batch.",
 				`this.type == "end_of_transaction"`,
-			).HasDefault(""),
+			).HasDefault("").Linter(docs.LintBloblangMapping),
 			docs.FieldDeprecated("condition").HasType(docs.FieldCondition).OmitWhen(func(v, _ interface{}) (string, bool) {
 				m, ok := v.(map[string]interface{})
 				if !ok {