+++ lib/processor/switch.go
@@ -51,7 +51,7 @@ At the end of switch processing the resulting batch will follow the same orderin
 				"A [Bloblang query](/docs/guides/bloblang/about/) that should return a boolean value indicating whether a message should have the processors of this case executed on it. If left empty the case always passes. If the check mapping throws an error the message will be flagged [as having failed](/docs/configuration/error_handling) and will not be tested against any other cases.",
 				`this.type == "foo"`,
 				`this.contents.urls.contains("https://benthos.dev/")`,
-			).HasDefault(""),
+			).HasDefault("").Linter(docs.LintBloblangMapping),
 			docs.FieldCommon(
 				"processors",
 				"A list of [processors](/docs/components/processors/about/) to execute on a message.",