+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Field `codec` added to the `socket` output.
 - New Bloblang method `map_each_key`.
 - General config linting improvements
+- Bloblang mappings and interpolated fields within configs are now compile checked during linting
 
 ### Fixed